splitting cda_identity into multiple files
